Add security to the views

+++ classes/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
-
+from django.contrib.auth.decorators import login_required
 from .models import Classes
 from .forms import ClassesForm
 
@@ -19,8 +19,13 @@ def classes(request):
     return render(request, 'classes/fitness_classes.html', context)
 
 
+@login_required
 def add_class(request):
-    """ Add a classes to the store """
+    """ Add a classes to the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only administrator can see this site.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ClassesForm(request.POST, request.FILES)
         if form.is_valid():
@@ -40,8 +45,13 @@ def add_class(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_class(request, classes_id):
-    """ Edit a class in the store """
+    """ Edit a class in the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only administrator can see this site.')
+        return redirect(reverse('home'))
+
     classes = get_object_or_404(Classes, pk=classes_id)
     if request.method == 'POST':
         form = ClassesForm(request.POST, request.FILES, instance=classes)
@@ -64,8 +74,13 @@ def edit_class(request, classes_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_class(request, classes_id):
-    """ Delete a class from the store """
+    """ Delete a class from the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only administrator can see this site.')
+        return redirect(reverse('home'))
+
     classes = get_object_or_404(Classes, pk=classes_id)
     classes.delete()
     messages.success(request, 'Class deleted!')
